requirements.txt
flask
pydantic
mysql-connector-python
PyJWT
Flask-Caching
redis
prometheus-flask-exporter
kafka-python
flake8
